requirements.txt
# Core Framework
numpy>=1.24.0
scipy>=1.10.0
opencv-python>=4.8.0
Pillow>=10.0.0

# ROS2 and Robotics
rclpy>=3.3.0
geometry-msgs>=4.2.0
sensor-msgs>=4.2.0
std-msgs>=4.2.0

# Machine Learning & AI
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0
scikit-learn>=1.3.0

# Computer Vision
mediapipe>=0.10.0
imutils>=0.5.4

# Hardware Interface
RPi.GPIO>=0.7.1
picamera2>=0.3.12
pyserial>=3.5
smbus2>=0.4.2
adafruit-circuitpython-motor>=3.4.0
adafruit-circuitpython-servo>=1.15.0

# Database and Storage
sqlite3
h5py>=3.9.0
pickle5>=0.0.11

# Communication
paho-mqtt>=1.6.0
flask>=2.3.0
flask-socketio>=5.3.0
websockets>=11.0.0

# Utilities
pyyaml>=6.0
configparser>=5.3.0
psutil>=5.9.0
matplotlib>=3.7.0
tqdm>=4.65.0
colorama>=0.4.6

# Development and Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Audio (optional)
pygame>=2.5.0
pydub>=0.25.0